Updates menu
This uses an HTMX out-of-band (oob) swap to update the menu when rendering a new page.

I recommend doing this a different way (client-side) since the menu may have security concerns that are both costly and don't change frequently. It is included here just to demonstrate doing this with HTMX without changing much code