Refactors Weather component with HtmxView
Makes the weather component work as intended with HTMX using the view.

This is obviously not how you'd want to do it in reality - you wouldn't get your data on the initial request and discard it only to do it again. But it demonstrates how you could differentiate between content of a menu navigation (boosted) vs the actual data load (non-boosted).

Also I'm not a fan of magic strings for URLs